refactor ovnmainlogic
This is a big commit that changes the implementation of the main logic,
ovnmonitor and also creates a wrapper for the iomodules.

ovnmonitor:
Now the ovnmonitor keeps a local copy (in an own format) of the monitored DB's
contents.  When an element inside a DB is changed, added or deleted, it first
parses this action to the own data format and then it notifies an upper layer
(mainlogic in this case) about the change.

wrapper for iomodules:
An IOModule is controlled through an instance of a class that represents
the actions that can be performed on the IOModule: create, destroy, attach
and detach ports for now.  This wrapper uses the already existing
implementation of hover control to send commands to the hover deamon on the
host.  Please note that this could be changed in the future, it should not
affect the global architecture as each hover is used only in the iomodules
classes.

mainlogic:
Main logic is a module that sits between ovnmonitor and the iomodules, it
processes notifications from the ovnmonitor and based on some internal
decisiones it issues the appropiate commands to the iomodules.

At this commit some features are not already available:
 - CLI
 - port security policies