iomodules/router: use lpm_trie map
Due to old eBPF limitations the LPM algorithm was done using a loop, in
the kernel v4.11 it has been introduced an implementation of an LPM map.
This commit uses the lpm_trie map to handle the routing table of the
router, this allows to have a larger router table and also to avoid the
needed to introduce the elements in a sorted way.